Various updates
Added issuedate, expirationdate, and evidence fields to the bulkedit form
Added validation function for bulkedit form
Added processing of above new fields when adding badges

Added new admin settings
 - issuer origin (required)
 - made default issuer name required
 - URL settings for baking and sending to the OBI backpack
 - Email notification values

Changed privateurl (assertion) to be indexed in the DB

Function to process requests for assertions and menu entry to route
the requests to new function.

Stub functions for baking a badge and notifying recipients.