Re-bake badges to send to backpack when user changes sharing option
Badges in the baked but not shared state should be re-baked when/if a
user changes their sharing option from "undecided" or "no" to "yes".

Add a function to do this, call it from the cron routine.

Change the baking routine to take a new parameter to indicate whether
the entry should be deleted upon a baking failure.  (Do not delete
when re-baking fails!)