Fix recipients form to include users with only revoked badges
Since we skip all the entries of revoked badges, if a user had
only revoked badges, they weren't included in the list of users.
Fix that.

Cosmetic: add space before uid in the list.

The db_rewrite_sql() is causing issues in the open_badges_notify_new_recipients()
function.  Don't use it for now...