Create views of current recipients
Either recipients of a certain badge, or all badges.

Also limit what is displayed to non-badge admin users to what
has been marked as allowed to display.

To include the display on an exisiting page, include the following:

<?php
  print open_badges_show_recipients(<bid>);
?>

Where <bid> is the badge id number of the particular badge,
or 'all' if all badges should be shown.